Changes related to chaining
phNxpEseProto7816_SendIframe -> the framebuffer is now a local buffer, which copies the relevant data from iFrameData.p_data and adds the crc and header. This way the data in iFrameData.p_data is not getting changed, which makes it better for handling chaining because we have to acess it later again to transmit all data.

IFSC_SIZE_SEND - limited the information field, so that data+header+crc are less than the 255 byte limit of the wire.h

changed many uint8_t variables to uint32_t , so it doesn't artificially limit the transmission to 255 byte (even with chainig)